fix signout button to the right, show order as all,delivered,undelevered and search by name of orderid

+++ app/controllers/categories_controller.rb
@@ -1,48 +1,79 @@
 class CategoriesController < ApplicationController
   before_filter :authenticate_user!
+  
   def new
-    @category=Category.new
+    if current_user.is_admin
+      @category = Category.new
+    else
+      flash[:error] = "Access denied"
+      redirect_to root_path
+    end  
   end
   
   def index
     @categories = Category.all
   end
+  
   def show
-    
+    @category = Category.find(params[:id])
+    @products = @category.products
   end
   
   def create
-    @category=Category.new(params[:category])
-    if @category.save
-      flash[:notice]="category is created successfully"  
-      redirect_to welcome_users_path
+    if current_user.is_admin
+      @category = Category.new(params[:category])
+      if @category.save
+        flash[:notice] = "category is created successfully" 
+        #TODO Redirect to index
+        redirect_to categories_path
+      else
+        flash[:partial] = "please enter category name"
+        redirect_to new_category_path  
+      end
     else
-      flash[:partial]="please enter category name"
-      redirect_to new_category_path  
-    end
+      flash[:error] = "Access denied"
+      redirect_to root_path
+    end  
   end
   
   def destroy
-    @category = Category.find(params[:id])
-    if @category.destroy
-      flash[:notice]="category remove"    
-      redirect_to categories_path
+    if current_user.is_admin
+      @category = Category.find(params[:id])
+      if @category.destroy
+        flash[:notice] = "category remove"  
+        redirect_to categories_path
+      else
+        flash[:partial] = "category is not remove"  
+      end
     else
-      flash[:partial]="category is not remove"  
-    end
+      flash[:error] = "Access denied"
+      redirect_to root_path
+    end  
   end
   
   def edit
-    @category=Category.find(params[:id])  
+    if current_user.is_admin
+      @category = Category.find(params[:id])  
+    else
+      flash[:error] = "Access denied"
+      redirect_to root_path
+    end
   end
   
   def update
-    @category = Category.find(params[:id])
-    if @category.update_attributes(params[:category])
-      flash[:notice]="category updated"
-      redirect_to welcome_users_path
+    if current_user.is_admin
+      @category = Category.find(params[:id])
+      if @category.update_attributes(params[:category])
+        flash[:notice] = "category updated"
+        #TODO Redirect to index
+        redirect_to categories_path
+      else
+        flash[:partial] = "category is not updated"  
+      end
     else
-      flash[:partial]="category is not updated"  
-    end
+      flash[:error] = "Access denied"
+      redirect_to root_path
+    end  
   end
+  
 end
